+++ src/video/scalers/Scale2xScaler.cc
@@ -249,7 +249,7 @@ void Scale2xScaler<Pixel>::scaleLineHalf_1on2(
 	dst[1] = (right == src0[0] && src2[0] != src0[0]) ? src0[0] : mid;
 
 	// Central pixels.
-	for (unsigned x = 1; x < srcWidth - 1; ++x) {
+	for (size_t x = 1; x < srcWidth - 1; ++x) {
 		Pixel left = mid;
 		mid   = right;
 		right = src1[x + 1];
@@ -297,7 +297,7 @@ void Scale2xScaler<Pixel>::scaleLineHalf_1on1(
 	dst[0] = mid;
 
 	// Central pixels.
-	for (unsigned x = 1; x < srcWidth - 1; ++x) {
+	for (size_t x = 1; x < srcWidth - 1; ++x) {
 		Pixel left = mid;
 		mid   = right;
 		right = src1[x + 1];
